# docs/management-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CodeSearch Management API",
    "version": "0.1.0",
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "description": "Local management server (default bind). Use 0.0.0.0 when started with --public.",
      "variables": {
        "port": {
          "default": "8676",
          "description": "Management port passed to `codesearch serve --mgmt-port` (default 8676)."
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/memory": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "List stored memory items",
        "operationId": "memoryList",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "description": "Restrict to one memory kind.",
            "schema": {
              "type": "string",
              "enum": [
                "preference",
                "experience",
                "skill",
                "fact"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stored memory items.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "count",
                    "items"
                  ],
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MemoryItem"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown memory kind.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/memory/search": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Search stored memories",
        "description": "Hybrid semantic + keyword search. Each result carries a `score` alongside the memory-item fields.",
        "operationId": "memorySearch",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "description": "Search query.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "num",
            "in": "query",
            "required": false,
            "description": "Maximum number of results.",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "description": "Restrict to one memory kind.",
            "schema": {
              "type": "string",
              "enum": [
                "preference",
                "experience",
                "skill",
                "fact"
              ]
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "description": "Restrict to memories relevant in this project/namespace scope (its items plus globals). Omit to search every scope.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scored memory results.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "count",
                    "results"
                  ],
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/MemoryItem"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "score": {
                                "type": "number",
                                "format": "float"
                              }
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown memory kind.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/memory/tree": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Browse the memory virtual filesystem",
        "description": "With no `uri`, returns the rollup node plus the sessions and resources directories. With a `uri`, returns that directory's children.",
        "operationId": "memoryTree",
        "parameters": [
          {
            "name": "uri",
            "in": "query",
            "required": false,
            "description": "Directory URI to list (e.g. `memory://sessions`).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nodes under the requested URI.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "count",
                    "nodes"
                  ],
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "nodes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MemoryNode"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/memory/dream": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Dream scheduler status",
        "description": "Scheduler configuration (from the `memory` section of `config.json`), whether a cycle is currently running, and the last recorded run. Returns `503` when the server started without a usable LLM backend.",
        "operationId": "memoryDreamStatus",
        "responses": {
          "200": {
            "description": "Scheduler status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "enabled",
                    "interval_hours",
                    "session_idle_minutes",
                    "auto_import",
                    "running"
                  ],
                  "properties": {
                    "enabled": {
                      "type": "boolean"
                    },
                    "interval_hours": {
                      "type": "integer"
                    },
                    "session_idle_minutes": {
                      "type": "integer"
                    },
                    "auto_import": {
                      "type": "boolean"
                    },
                    "running": {
                      "type": "boolean"
                    },
                    "last_run": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/DreamRun"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Dreaming is not available on this server."
          }
        }
      },
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Trigger a dream cycle",
        "description": "Starts a dream cycle (harvest finished sessions, consolidate near-duplicate/contradictory memories, reflect, refresh the rollup) in the background and returns immediately. Poll `GET /api/memory/dream` for the recorded outcome.",
        "operationId": "memoryDreamTrigger",
        "responses": {
          "202": {
            "description": "Cycle started.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "started"
                  ],
                  "properties": {
                    "started": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "A dream cycle is already running."
          },
          "503": {
            "description": "Dreaming is not available on this server."
          }
        }
      }
    },
    "/api/memory/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "A memory item UUID, a `kind/name` reference, or a `memory://\u2026` node URI.",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Get one memory item or virtual-filesystem node",
        "description": "Returns `{ \"item\": MemoryItem }` for an item address, or `{ \"node\": MemoryNode }` when the id is a `memory://` URI.",
        "operationId": "memoryGet",
        "responses": {
          "200": {
            "description": "The resolved item or node.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "item"
                      ],
                      "properties": {
                        "item": {
                          "$ref": "#/components/schemas/MemoryItem"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "node"
                      ],
                      "properties": {
                        "node": {
                          "$ref": "#/components/schemas/MemoryNode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No item or node at the given address.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Uniform error body returned with a 4xx/5xx status.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable failure message."
          }
        }
      },
...
      "MemoryItem": {
        "type": "object",
        "description": "Serialized memory item: a durable fact/preference/experience/skill extracted from a session.",
        "additionalProperties": true
      },
      "MemoryNode": {
        "type": "object",
        "description": "Serialized memory virtual-filesystem node (rollup, session, or resource directory/leaf).",
        "additionalProperties": true
      },
...
      "DreamRun": {
        "type": "object",
        "description": "Record of one completed dream (offline memory consolidation) cycle.",
        "required": [
          "id",
          "started_at",
          "finished_at",
          "sessions_imported",
          "clusters_found",
          "operations_applied",
          "operations_skipped"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "started_at": {
            "type": "integer",
            "description": "Unix seconds."
          },
          "finished_at": {
            "type": "integer",
            "description": "Unix seconds."
          },
          "sessions_imported": {
            "type": "integer"
          },
          "clusters_found": {
            "type": "integer"
          },
          "operations_applied": {
            "type": "integer"
          },
          "operations_skipped": {
            "type": "integer"
          }
        }
      }
    }
  }
}